+++ pymadagascar/localization/__init__.py
@@ -3,19 +3,23 @@
 from .traveltime import (
     LocalizationError,
     LocalizationGridSearchResult,
+    VariableVelocityLocalizationGridSearchResult,
     diffraction_travel_time_2d,
     direct_travel_time_2d,
     euclidean_distance_2d,
     grid_search_point_location_2d,
+    grid_search_point_location_velocity_2d,
     travel_time_residuals,
 )
 
 __all__ = [
     "LocalizationError",
     "LocalizationGridSearchResult",
+    "VariableVelocityLocalizationGridSearchResult",
     "diffraction_travel_time_2d",
     "direct_travel_time_2d",
     "euclidean_distance_2d",
     "grid_search_point_location_2d",
+    "grid_search_point_location_velocity_2d",
     "travel_time_residuals",
 ]